pyproject.toml
[tool.poetry]
name = "agency_swarm_project"
version = "0.1.0"
description = "Um projeto para criar e gerenciar campanhas de anúncios utilizando o Agency Swarm Tool."
authors = ["Anderson Vieira <andersonvilarins@gmail.com>"]

[tool.poetry.dependencies]
agency-swarm = "0.1.6"
python = ">=3.10,<4.0"
python-dotenv = "1.0.0"
"facebook-business" = "^12.0"
requests = "^2.31"
pydantic = ">=2.0.2,<3.0.0"
gradio = "^4.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
